Chore | Extract K8s client into dedicated pkg/k8s package
Move all Kubernetes client functionality from pkg/utils into a new
pkg/k8s package, splitting the monolithic k8s-client.go (1024 lines)
into focused files: client.go, manifest.go, argocd.go, discovery.go,
deployment.go, portforward.go, and crash_watcher.go.

Renames utils.K8sClient to k8s.Client and utils.NewK8sClient to
k8s.NewClient. The utils package retains only general-purpose helpers
(WriteFile, CreateFolder, UniqueId, SplitYAMLDocuments).
