+++ pkg/argocd/operations.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/dag-andersen/argocd-diff-preview/pkg/utils"
+	"github.com/dag-andersen/argocd-diff-preview/pkg/k8s"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/vars"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
@@ -65,7 +65,7 @@ type apiConnection struct {
 // If renderMode is cli, returns a CLIOperations instance.
 // The authToken parameter is optional - if provided, it will be used instead of
 // authenticating with the ArgoCD server during Login().
-func NewOperations(renderMode vars.RenderMethod, k8sClient *utils.K8sClient, namespace string, loginOptions string, authToken string) Operations {
+func NewOperations(renderMode vars.RenderMethod, k8sClient *k8s.Client, namespace string, loginOptions string, authToken string) Operations {
 	switch renderMode {
 	case vars.RenderMethodServerAPI, vars.RenderMethodRepoServerAPI:
 		return &APIOperations{
